pyproject.toml
[project]
name = "snmedical"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "langchain",
    "langchain-community",
    "langchain-core",
    "langchain-groq",
    "pinecone",
    "langchain-pinecone",
    "langchain-google-genai",
    "PyPDF",
    "python-dotenv",
    "pydantic",
    "requests",
    "tqdm",
    "loguru",
    "streamlit",
    "requests",
]
